requirements.txt
```text
python-dotenv==1.0.0
langchain==0.1.0
langchain-core==0.1.8
langchain-openai==0.0.2.post1
langgraph==0.0.20
pydantic==2.4.2
openai==1.3.7
streamlit==1.28.0
langsmith==0.0.83
colorama==0.4.6